Return the call to vuex.store.dispatch
Vuex actions are typically asynchronous and, therefore, benefit from being able to return promises. The dispatch helper needs to return the dispatch call so that any promise that is returned gets returned to the caller. The commit helper should probably have it's call result returned as well, but the use case is atypical.